Added files to test code using Moq 4 and xUnit

+++ ICalculator.cs
@@ -8,9 +8,9 @@ namespace Calculator
 {
     internal interface ICalculator
     {
-        int add(int a, int b);
-        int subtract(int a, int b);
-        int multiply(int a, int b);
-        double divide(int a, int b);
+        double add(double a, double b);
+        double subtract(double a, double b);
+        double multiply(double a, double b);
+        double divide(double a, double b);
     }
 }
